Componized the result table to be able to accept headers and generic items as props. Now the table emits an event when the row is clicked.
