+++ src/components/admin/ResultTable.vue
@@ -2,8 +2,6 @@
   import { computed, type ComputedRef } from 'vue'
   import LayoutCard from '@/components/cards/LayoutCard.vue'
   import { type Composer, useI18n } from 'vue-i18n'
-  import { type Personendatensatz } from '@/stores/PersonStore'
-  import { type Router, useRouter } from 'vue-router'
 
   /* this block is necessary to introduce a table header type for defining table headers
       watch source for updates: https://stackoverflow.com/a/75993081/4790594
@@ -11,25 +9,17 @@
   import type { VDataTableServer } from 'vuetify/lib/components/index.mjs'
   type ReadonlyHeaders = InstanceType<typeof VDataTableServer>['headers']
 
+  type GenericItem = Record<string, unknown>
+
   type Props = {
-    items: Personendatensatz[]
+    items: GenericItem[]
     loading: boolean
     totalItems: number
+    headers: ReadonlyHeaders
   }
-
   const props: Props = defineProps<Props>()
 
   const { t }: Composer = useI18n({ useScope: 'global' })
-  const router: Router = useRouter()
-
-  const headers: ReadonlyHeaders = [
-    { title: t('user.lastName'), key: 'person.name.familienname', align: 'start' },
-    { title: t('user.firstName'), key: 'person.name.vorname', align: 'start' }
-  ]
-
-  function handleRowClick(_$event: PointerEvent, { item }: { item: Personendatensatz }): void {
-    router.push({ name: 'user-details', params: { id: item.person.id } })
-  }
 
   // TODO: these two values will come from the API in the future
   const itemsPerPage: number = 25
@@ -49,7 +39,9 @@
   <LayoutCard :header="$t('admin.user.management')">
     <v-data-table-server
       class="user-table"
-      @click:row="handleRowClick"
+      @click:row="
+        ($event: PointerEvent, item: GenericItem[]) => $emit('onHandleRowClick', $event, item)
+      "
       data-testid="user-table"
       density="compact"
       :headers="headers"
